missed files on previous commit. Gets character grids showing all data correctly and adding css classes appropriately

+++ Models/BlizzardApiModels/RatingInfo.cs
@@ -3,6 +3,6 @@ namespace Singularity.Models.BlizzardApiModels
     public class RatingInfo
     {
         public Color Color { get; set; } = default!;
-        public double Rating { get; set; } = default!;
+        public decimal Rating { get; set; } = default!;
     }
 }
